+++ crates/jlrs/tests/derive.rs
@@ -1014,6 +1014,62 @@ mod tests {
         })
     }
 
+    fn with_u128() {
+        JULIA_DERIVE.with(|handle| {
+            handle.borrow().local_scope::<_, 4>(|mut frame| {
+                let a = -1;
+                let b = 0xffffffffffffffff1;
+                let c = -2;
+
+                let with_u128 = WithU128::new(a, b, c);
+                let with_u128_v = Value::new(&mut frame, with_u128);
+                assert_eq!(with_u128_v.datatype().size().unwrap() as usize, size_of::<WithU128>());
+
+                let a_v = with_u128_v.get_field(&mut frame, "a").unwrap();
+                assert_eq!(a_v.unbox::<i8>().unwrap(), a);
+
+                let b_v = with_u128_v.get_field(&mut frame, "b").unwrap();
+                assert_eq!(b_v.unbox::<u128>().unwrap(), b);
+
+                let c_v = with_u128_v.get_field(&mut frame, "c").unwrap();
+                assert_eq!(c_v.unbox::<i8>().unwrap(), c);
+
+                let unboxed = with_u128_v.unbox::<WithU128>().unwrap();
+                assert_eq!(unboxed.a(), a);
+                assert_eq!(unboxed.b(), b);
+                assert_eq!(unboxed.c(), c);
+            });
+        })
+    }
+
+    fn with_i128() {
+        JULIA_DERIVE.with(|handle| {
+            handle.borrow().local_scope::<_, 4>(|mut frame| {
+                let a = -1;
+                let b = -0xffffffffffffffff1;
+                let c = -2;
+
+                let with_i128 = WithI128::new(a, b, c);
+                let with_i128_v = Value::new(&mut frame, with_i128);
+                assert_eq!(with_i128_v.datatype().size().unwrap() as usize, size_of::<WithI128>());
+
+                let a_v = with_i128_v.get_field(&mut frame, "a").unwrap();
+                assert_eq!(a_v.unbox::<i8>().unwrap(), a);
+
+                let b_v = with_i128_v.get_field(&mut frame, "b").unwrap();
+                assert_eq!(b_v.unbox::<i128>().unwrap(), b);
+
+                let c_v = with_i128_v.get_field(&mut frame, "c").unwrap();
+                assert_eq!(c_v.unbox::<i8>().unwrap(), c);
+
+                let unboxed = with_i128_v.unbox::<WithI128>().unwrap();
+                assert_eq!(unboxed.a(), a);
+                assert_eq!(unboxed.b(), b);
+                assert_eq!(unboxed.c(), c);
+            });
+        })
+    }
+
     fn test_enums() {
         JULIA_DERIVE.with(|handle| {
             handle.borrow_mut().with_stack(|mut stack| {
@@ -1147,5 +1203,7 @@ mod tests {
         test_enums();
         test_enums_ccall();
         derive_complex_bits_union();
+        with_u128();
+        with_i128();
     }
 }